Create re_build_web_viewer
