+++ Cargo.toml
@@ -27,6 +27,7 @@ re_analytics = { path = "crates/re_analytics", version = "0.2.0" }
 re_arrow_store = { path = "crates/re_arrow_store", version = "0.2.0" }
 re_build_build_info = { path = "crates/re_build_build_info", version = "0.2.0" }
 re_build_info = { path = "crates/re_build_info", version = "0.2.0" }
+re_build_web_viewer = { path = "crates/re_build_web_viewer", version = "0.2.0" }
 re_data_store = { path = "crates/re_data_store", version = "0.2.0" }
 re_error = { path = "crates/re_error", version = "0.2.0" }
 re_format = { path = "crates/re_format", version = "0.2.0" }
